feat(quiz): ms-precision scoring, shuffle answer options, leaderboard tie-breaks
- scoring: measure answer time via Stopwatch in milliseconds instead of whole
  seconds (3.1s and 3.9s now yield different points); totalElapsedSeconds → double
- questions: shuffle answer options per round with correctIndex remap
  (index-permutation, dup-safe); true/false kept in fixed order
- leaderboard: tie-break chain score↓ → avg_seconds↑ → updated_at↑ → user_id

+++ lib/src/bloc/quiz_state.dart
@@ -23,7 +23,7 @@ class QuizGameState extends Equatable {
     this.score = 0,
     this.correctCount = 0,
     this.lastEarnedScore = 0,
-    this.totalElapsedSeconds = 0,
+    this.totalElapsedSeconds = 0.0,
     this.activeCategoryId,
     this.weeklyLeaderboard = const [],
     this.isLoadingLeaderboard = false,
@@ -49,9 +49,9 @@ class QuizGameState extends Equatable {
   /// Score earned on the last answer (for +N pill in toast).
   final int lastEarnedScore;
 
-  /// Sum of (questionSeconds - secondsLeft) across answered questions — used
-  /// for the "average speed" stat on results.
-  final int totalElapsedSeconds;
+  /// Sum of precise elapsed time (в секундах, с долями) across answered
+  /// questions — used for the "average speed" stat on results.
+  final double totalElapsedSeconds;
 
   /// `null` between rounds.
   final String? activeCategoryId;
@@ -98,7 +98,7 @@ class QuizGameState extends Equatable {
     int? score,
     int? correctCount,
     int? lastEarnedScore,
-    int? totalElapsedSeconds,
+    double? totalElapsedSeconds,
     String? Function()? activeCategoryId,
     List<QuizLeaderboardEntry>? weeklyLeaderboard,
     bool? isLoadingLeaderboard,
